+++ MiHoYoUID/const.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-PLUGIN_NAME = "gscore_miao-plugin"
+PLUGIN_NAME = "MiHoYoUID"
 DATA_DIR_NAME = "GsCoreMiao"
 
 PACKAGE_DIR = Path(__file__).resolve().parent